services/tickets/requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
qrcode>=7.4.2
pillow>=11.0.0
python-dotenv>=1.0.0
httpx>=0.24.1
email-validator>=2.0.0
psycopg2-binary>=2.9.9
